Wfh/add semistructured (#38)
Refactoring to share rag things

---------
