+++ langchain-docs-benchmarking/packages/openai-functions-agent/openai_functions_agent/agent.py
