+++ langchain_benchmarks/extraction/email_task.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from typing import Optional, List
+from typing import List, Optional
 
 from langchain.smith import RunEvalConfig
 from pydantic import BaseModel, Field
