+++ langchain_benchmarks/rag/.gitignore
@@ -1,0 +1,1 @@
+*.sql
