+++ langchain_benchmarks/rag/__init__.py
@@ -1,8 +1,5 @@
-"""RAG environments."""
-from langchain_benchmarks.rag.evaluators import RAG_EVALUATION
-from langchain_benchmarks.rag.environments.langchain_docs.task import (
-    LANGCHAIN_DOCS_TASK,
-)
+from langchain_benchmarks.rag.evaluators import get_eval_config
+from langchain_benchmarks.rag.tasks import LANGCHAIN_DOCS_TASK
 
-# Please keep this list sorted!
-__all__ = ["LANGCHAIN_DOCS_TASK", "RAG_EVALUATION"]
+# Please keep this sorted
+__all__ = ["get_eval_config", "LANGCHAIN_DOCS_TASK"]
