--- langchain_benchmarks/rag/environments/langchain_docs/README.md
@@ -1,6 +1,0 @@
-# LangChain Docs Environment
-
-This code contains utilities to scrape the LangChain docs (already run) and index them
-using common techniques. The docs were scraped using the code in `_ingest_docs.py` and
-uploaded to gcs. To better compare retrieval techniques, we hold these constant and pull
-from that cache whenever generating different indices.
