--- langchain_benchmarks/rag/environments/langchain_docs/architectures/__init__.py
@@ -1,5 +1,0 @@
-from langchain_benchmarks.rag.environments.langchain_docs.architectures.chain_registry import (
-    ARCH_FACTORIES,
-)
-
-__all__ = ["ARCH_FACTORIES"]
