--- langchain_benchmarks/rag/environments/langchain_docs/langchain_docs_retriever/.gitignore
@@ -1,4 +1,0 @@
-db/
-docs_db/
-.sql
-.bin
