--- langchain_benchmarks/rag/environments/langchain_docs/langchain_docs_retriever/__init__.py
@@ -1,18 +1,0 @@
-from langchain_benchmarks.rag.environments.langchain_docs.langchain_docs_retriever.retriever import (
-    get_parent_document_retriever,
-    get_vectorstore_retriever,
-    get_hyde_retriever,
-    create_index,
-)
-from langchain_benchmarks.rag.environments.langchain_docs.langchain_docs_retriever.retriever_registry import (
-    RETRIEVER_FACTORIES,
-)
-
-__all__ = [
-    "create_index",
-    "get_hyde_retriever",
-    "get_parent_document_retriever",
-    "get_retriever",
-    "get_vectorstore_retriever",
-    "RETRIEVER_FACTORIES",
-]
