+++ langchain_benchmarks/rag/tasks/__init__.py
@@ -1,0 +1,7 @@
+from langchain_benchmarks.rag.tasks.langchain_docs.task import LANGCHAIN_DOCS_TASK
+from langchain_benchmarks.rag.tasks.semi_structured_earnings.task import (
+    SEMI_STRUCTURED_EARNINGS_TASK,
+)
+
+# Please keep this sorted
+__all__ = ["LANGCHAIN_DOCS_TASK", "SEMI_STRUCTURED_EARNINGS_TASK"]
