+++ langchain_benchmarks/rag/tasks/langchain_docs/__init__.py
@@ -1,0 +1,8 @@
+from langchain_benchmarks.rag.tasks.langchain_docs import architectures, indexing
+from langchain_benchmarks.rag.tasks.langchain_docs.task import LANGCHAIN_DOCS_TASK
+
+DATASET_ID = (
+    "452ccafc-18e1-4314-885b-edd735f17b9d"  # ID of public LangChain Docs dataset
+)
+
+__all__ = ["architectures", "indexing", "DATASET_ID", "LANGCHAIN_DOCS_TASK"]
