+++ langchain_benchmarks/rag/tasks/langchain_docs/architectures/__init__.py
@@ -1,0 +1,5 @@
+from langchain_benchmarks.rag.tasks.langchain_docs.architectures.chain_registry import (
+    ARCH_FACTORIES,
+)
+
+__all__ = ["ARCH_FACTORIES"]
