+++ langchain_benchmarks/rag/tasks/langchain_docs/indexing/.gitignore
@@ -1,0 +1,4 @@
+db/
+db_docs/
+.sql
+.bin
