+++ langchain_benchmarks/rag/tasks/langchain_docs/indexing/__init__.py
@@ -1,0 +1,7 @@
+from langchain_benchmarks.rag.tasks.langchain_docs.indexing.retriever_registry import (
+    RETRIEVER_FACTORIES,
+)
+
+__all__ = [
+    "RETRIEVER_FACTORIES",
+]
