+++ langchain_benchmarks/rag/tasks/semi_structured_earnings/indexing/.gitignore
@@ -1,0 +1,1 @@
+pdfs/
