+++ langchain_benchmarks/rag/tasks/semi_structured_earnings/indexing/__init__.py
@@ -1,0 +1,6 @@
+from langchain_benchmarks.rag.tasks.semi_structured_earnings.indexing.retriever_registry import (
+    RETRIEVER_FACTORIES,
+    load_docs,
+)
+
+__all__ = ["RETRIEVER_FACTORIES", "load_docs"]
