+++ langchain_benchmarks/registration.py
@@ -1,13 +1,16 @@
 """Registry of environments for ease of access."""
 
 from langchain_benchmarks.extraction import email_task
+from langchain_benchmarks.rag.tasks import (
+    LANGCHAIN_DOCS_TASK,
+    SEMI_STRUCTURED_EARNINGS_TASK,
+)
 from langchain_benchmarks.schema import Registry
-from langchain_benchmarks.rag import LANGCHAIN_DOCS_TASK
 from langchain_benchmarks.tool_usage import (
-    type_writer_26_funcs,
-    type_writer,
-    relational_data,
     multiverse_math,
+    relational_data,
+    type_writer,
+    type_writer_26_funcs,
 )
 
 # Using lower case naming to make a bit prettier API when used in a notebook
@@ -19,5 +22,6 @@ registry = Registry(
         multiverse_math.MULTIVERSE_MATH,
         email_task.EMAIL_EXTRACTION_TASK,
         LANGCHAIN_DOCS_TASK,
+        SEMI_STRUCTURED_EARNINGS_TASK,
     ]
 )
